+++ datumaro/cli/commands/__init__.py
@@ -24,6 +24,7 @@ from . import (
     merge,
     patch,
     remove,
+    search,
     stats,
     status,
     transform,